chore(I18n): move i18n context into new 'context' dir
