+++ .storybook/preview.js
@@ -3,7 +3,8 @@ import { ThemeProvider as EmotionThemeProvider } from 'emotion-theming'
 import { ThemeProvider } from '@mui/material/styles'
 import { CssBaseline } from '@mui/material'
 import { settings } from '~/api/__mock__'
-import { CheckoutProvider, I18nProvider, SettingsProvider } from '~/api'
+import { CheckoutProvider, SettingsProvider } from '~/api'
+import { I18nProvider } from '~/context'
 import { AppProvider } from '~/containers/App/AppContext'
 import { createTheme } from '~/components/styles'
 import breakpoints from '~/components/styles/breakpoints'
