+++ README.md
@@ -61,7 +61,7 @@ The philosophy behind some of this structure can be read below.
 
 ### src/api
 
-This is a good place to add App-level providers that should be accessible from all corners of the application. Providers added here should be exported in the api directory's common export file(s). Accessing data or helpers from these providers can look like this `import { useI18n } from '~/api'`.
+This is a good place to add App-level providers that should be accessible from all corners of the application. Providers added here should be exported in the api directory's common export file(s). Accessing data or helpers from these providers can look like this `import { useCms } from '~/api'`.
 
 ```
 src/api
