+++ src/api/index.js
@@ -1,5 +1,4 @@
 export * from './CheckoutContext'
-export * from './I18nContext'
 export * from './SettingsContext'
 
 /** @todo Replace mocked api calls with real implementation. */
