+++ src/api/storybook.index.js
@@ -2,7 +2,6 @@ import { sleep } from '@noaignite/utils'
 import { filter, products } from './__mock__'
 
 export * from './CheckoutContext'
-export * from './I18nContext'
 export * from './SettingsContext'
 
 export async function fetchProducts(query) {
