+++ src/blocks/ArticleSlideshow/ArticleSlideshow.js
@@ -4,7 +4,7 @@ import useEmblaCarousel from 'embla-carousel-react'
 import { styled } from '@mui/system'
 import { Button, Typography } from '@mui/material'
 import { Media, MediaReveal } from '@noaignite/oui'
-import { useI18n } from '~/api'
+import { useI18n } from '~/context'
 import { ASPECT_RATIOS } from '~/utils/constants'
 import { RouterLink } from '~/containers'
