+++ src/containers/App/partials/AppNavDrawer.js
@@ -4,7 +4,8 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 import { styled } from '@mui/system'
 import { Drawer, IconButton, Link, Toolbar, Typography } from '@mui/material'
-import { useCheckoutSelection, useI18n, useSettings } from '~/api'
+import { useCheckoutSelection, useSettings } from '~/api'
+import { useI18n } from '~/context'
 import { Close as CloseIcon } from '~/components/icons'
 import { useApp } from '../AppContext'
 import AppNavDrawerListItem from './AppNavDrawerListItem'
