+++ src/containers/CartItem/CartItem.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types'
 import { styled } from '@mui/system'
 import { ButtonBase, Link } from '@mui/material'
 import { Media, MediaReveal } from '@noaignite/oui'
-import { useCheckoutHandlers, useI18n } from '~/api'
+import { useCheckoutHandlers } from '~/api'
+import { useI18n } from '~/context'
 import { ASPECT_RATIOS, CENTRA_CART_ITEM_UNIQUE_KEY } from '~/utils/constants'
 import { cartItemType } from '~/utils'
 import { Add as AddIcon, Remove as RemoveIcon } from '~/components/icons'
