+++ src/containers/CartSummary/CartSummary.js
@@ -2,7 +2,8 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 import { styled } from '@mui/system'
 import { Typography } from '@mui/material'
-import { useCheckoutSelection, useI18n } from '~/api'
+import { useCheckoutSelection } from '~/api'
+import { useI18n } from '~/context'
 import { cartItemType } from '~/utils'
 
 const CartSummaryRoot = styled('div', {
