+++ src/pages/_app.js
@@ -8,7 +8,8 @@ import { CacheProvider } from '@emotion/react'
 import { ThemeProvider } from '@mui/material/styles'
 import { CssBaseline } from '@mui/material'
 import { settings as mockedCmsProps } from '~/api/__mock__'
-import { CheckoutProvider, I18nProvider, SettingsProvider } from '~/api'
+import { CheckoutProvider, SettingsProvider } from '~/api'
+import { I18nProvider } from '~/context'
 import createEmotionCache from '~/utils/createEmotionCache'
 import theme from '~/utils/theme.light'
 import { AppProvider } from '~/containers/App/AppContext'
